Populate address information

+++ InteliPharma.DB/App/Tables/Street.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [App].[Street] (
     [StreetId]       INT             IDENTITY (1, 1) NOT NULL,
-    [NeighborhoodId] SMALLINT        NOT NULL,
+    [NeighborhoodId] INT        NOT NULL,
     [StreetName]     VARCHAR (50)    NOT NULL,
     [Longitude]      DECIMAL (12, 9) NULL,
     [Latitude]       DECIMAL (12, 9) NULL,
